requirements.txt
streamlit
groq
python-docx
docx2txt
fpdf
graphviz
diff-match-patch